fixed various small issues, re-enabled old protocol, was not able to test on real hardware yet (but might work again)

+++ trunk/emuv1/msxtypes.h
@@ -8,7 +8,7 @@ typedef unsigned char nw_byte;          // should be (at least?) an 8 bit type
 typedef unsigned int nw_word;           // should be (at least?) an 16 bit type
     
 typedef unsigned int emuTimeType;
-typedef unsigned int msTimeType;
+typedef unsigned long long msTimeType;
 
 #endif
